Implement linked list data structure utility
